Feature/search4 customer (#9)
* commented

* used indentation

* service file for search feature

* added search component

* search button working: check console

* data sent from header component, ready to receive from search comp

* module added for form

* test search cmp

* search button working.
search when enter pressed

* insert data into search comp

* showing data  on data entered in search bar

* pagination working for search feature

* updated pagination

* Fix: Rendering inconsistency in app-search component layout

Added a wrapping <div> around the app-search component within the <ng-template> in order to resolve a rendering inconsistency that was affecting styling and visibility. This change ensures a consistent layout and styling behavior across the template. Issue resolved by addressing how the component was nested within the template structure.

---------

Co-authored-by: boro_vai <83267669+100getSetter@users.noreply.github.com>